string.cpp file added
File Details

Input Format
You are given two strings,  and , separated by a new line. Each string will consist of lower case Latin characters ('a'-'z').

Output Format
In the first line print two space-separated integers, representing the length of  and  respectively. 
In the second line print the string produced by concatenating  and  (). 
In the third line print two strings separated by a space,  and . and  are the same as  and , respectively, except that their first characters are swapped.